moved remaining repeated components into correct package
